refactor!: controllers & namespace refactoring
