+++ src/Controller/API/ProjectController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controller;
+namespace App\Controller\API;
 
 use App\Entity\Project;
 use App\Repository\ProjectRepository;
