+++ src/Controller/API/TechnologyController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controller;
+namespace App\Controller\API;
 
 use App\Entity\Technology;
 use App\Repository\TechnologyRepository;
